# Minor.Dag18.FrontEndOpdracht/src/MonumentenGeneratedPageAndAgent/Agents/Models/Monument.cs
// Code generated by Microsoft (R) AutoRest Code Generator 0.16.0.0
// Changes may cause incorrect behavior and will be lost if the code is
// regenerated.

namespace MonumentenGeneratedPageAndAgent.Models
{
    using System;
    using System.Linq;
    using System.Collections.Generic;
    using Newtonsoft.Json;
    using Microsoft.Rest;
    using Microsoft.Rest.Serialization;

    public partial class Monument
    {
        /// <summary>
        /// Initializes a new instance of the Monument class.
        /// </summary>
        public Monument() { }

        /// <summary>
        /// Initializes a new instance of the Monument class.
        /// </summary>
        public Monument(int? id = default(int?), string naam = default(string))
        {
            Id = id;
            Naam = naam;
        }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "id")]
        public int? Id { get; set; }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "naam")]
        public string Naam { get; set; }

    }
}